refactor: move atomic file write to shared layer
